cleargbm_rs: training loop — leaf-cache fast path, f32 grad/hess, u32 subsample indices
Rewires `libs/cleargbm_rs/src/training/train.rs` to call `build_tree_with_leaf_assignment` and take the direct fast path (`raw_preds_train[i] += lr * leaf_value_per_sample[i]`) for every row whose leaf value is not the NaN sentinel, only falling back to `predict_tree` on a collected list of subsampled-out indices — when `subsample = 1.0`, zero tree walks per round. Gradients and hessians are computed directly as `Vec<f32>` via `crate::narrow::score_narrow(p - f64::from(y))` and `score_narrow(p * (1.0 - p))`, so the histogram accumulator receives the narrowed `f32` scores at the input boundary and stays `f64` in the sums. `subsampling.rs::get_sample_indices` returns `Vec<u32>` (validates `n_samples > u32::MAX` at construction with `IntegerConversion`), and `rng.rs::shuffle_partial` takes `&mut [u32]` to match — the shuffle body is unchanged since `slice.swap` works generically over any element type.

+++ libs/cleargbm_rs/src/training/rng.rs
@@ -146,7 +146,7 @@ impl SimpleRng {
     /// Returns `ClearGbmError::InvalidParameter` if `n_take > slice.len()`.
     pub(crate) fn shuffle_partial(
         &mut self,
-        slice: &mut [usize],
+        slice: &mut [u32],
         n_take: usize,
     ) -> Result<(), ClearGbmError> {
         let len = slice.len();
